Update type checking to work python 3.8
